알림 시간 설정 적용

+++ src/main/java/jpcompany/smartwire/firebase/service/FCMNotificationService.java
@@ -1,9 +1,6 @@
 package jpcompany.smartwire.firebase.service;
 
-import com.google.firebase.messaging.FirebaseMessaging;
-import com.google.firebase.messaging.FirebaseMessagingException;
-import com.google.firebase.messaging.Message;
-import com.google.firebase.messaging.Notification;
+import com.google.firebase.messaging.*;
 import jpcompany.smartwire.domain.Member;
 import jpcompany.smartwire.firebase.dto.FCMNotificationDto;
 import jpcompany.smartwire.mobile.dto.FCMTokenAndAlarmSettingDto;
@@ -13,6 +10,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -25,7 +23,40 @@ public class FCMNotificationService {
     private final MemberRepository memberRepository;
     private final MobileRepository mobileRepository;
 
+    private final AndroidConfig androidConfig = AndroidConfig.builder()
+            .setPriority(AndroidConfig.Priority.HIGH) // 최대 우선순위 설정
+            .setNotification(AndroidNotification.builder()
+                    .setSound("default") // 소리 설정
+                    .build()
+            )
+            .build();
+
+    private final ApnsConfig apnsConfig = ApnsConfig.builder()
+            .putHeader("apns-priority", "10") // 최대 우선순위로 설정 (10은 최대값입니다)
+            .setAps(Aps.builder()
+                    .setSound("default") // 기본 소리를 설정하거나 원하는 소리 파일 이름을 제공할 수 있습니다.
+                    .build())
+            .build();
+
+    private boolean isTime(LocalTime now, String alarmSetting) {
+        String alarmStartTimeStr = alarmSetting.substring(1).split(",")[0];
+        String alarmLastTimeStr = alarmSetting.substring(1).split(",")[1];
+        LocalTime alarmStartTime = LocalTime.of(
+                Integer.parseInt(alarmStartTimeStr.split(":")[0]),
+                Integer.parseInt(alarmStartTimeStr.split(":")[1]), 0);
+        LocalTime alarmLastTime = LocalTime.of(
+                Integer.parseInt(alarmLastTimeStr.split(":")[0]),
+                Integer.parseInt(alarmLastTimeStr.split(":")[1]), 0);
+
+        if (alarmStartTime.isBefore(alarmLastTime)) {
+            return now.isAfter(alarmStartTime) && now.isBefore(alarmLastTime);
+        } else {
+            return now.isAfter(alarmStartTime) || now.isBefore(alarmLastTime);
+        }
+    }
+
     public void sendNotificationByToken(FCMNotificationDto notificationDto) {
+        LocalTime now = LocalTime.now();
         Optional<Member> member = memberRepository.findById(notificationDto.getTargetMemberId());
 
         if (member.isPresent()) {
@@ -37,32 +68,39 @@ public class FCMNotificationService {
                         .build();
 
                 for (FCMTokenAndAlarmSettingDto fcmTokenAndAlarmSettingDto : fcmTokenAndAlarmSettingDtoList) {
-                    Message message = Message.builder()
-                            .setToken(fcmTokenAndAlarmSettingDto.getFcmToken())
-                            .setNotification(notification)
-                            .build();
-                    try {
-                        firebaseMessaging.send(message);
+                    String alarmSetting = fcmTokenAndAlarmSettingDto.getAlarmSetting();
 
-                        log.info( "푸시 알림 성공 / id = {} / token = {}",
-                                notificationDto.getTargetMemberId(),
-                                fcmTokenAndAlarmSettingDto.getFcmToken()
-                        );
+                    if (alarmSetting.charAt(0) == 'y' && isTime(now, alarmSetting)) {
+                        Message message = Message.builder()
+                                .setToken(fcmTokenAndAlarmSettingDto.getFcmToken())
+                                .setNotification(notification)
+                                .setAndroidConfig(androidConfig)
+                                .setApnsConfig(apnsConfig)
+                                .build();
+                        try {
+                            firebaseMessaging.send(message);
+                            log.info( "푸시 알림 성공 / id = {} / token = {}",
+                                    notificationDto.getTargetMemberId(),
+                                    fcmTokenAndAlarmSettingDto.getFcmToken()
+                            );
 
-                    } catch (FirebaseMessagingException e) {
-                        if (e.getMessage().contains("not a valid FCM")) {
-                            mobileRepository.deleteFCMTokenById(fcmTokenAndAlarmSettingDto.getId());
-                            log.info("유효하지 않은 FCM Token 삭제 = {}", fcmTokenAndAlarmSettingDto.getFcmToken());
-                            return;
+                        } catch (FirebaseMessagingException e) {
+                            if (e.getMessage().contains("not a valid FCM")) {
+                                mobileRepository.deleteFCMTokenById(fcmTokenAndAlarmSettingDto.getId());
+                                log.info("유효하지 않은 FCM Token 삭제 = {}", fcmTokenAndAlarmSettingDto.getFcmToken());
+                                return;
+                            }
+
+                            log.error( "푸시 알림 실패 / id = {} / token = {} / message = {}",
+                                    notificationDto.getTargetMemberId(),
+                                    fcmTokenAndAlarmSettingDto.getFcmToken(),
+                                    e.getMessage()
+                            );
+
                         }
-
-                        log.error( "푸시 알림 실패 / id = {} / token = {} / message = {}",
-                                notificationDto.getTargetMemberId(),
-                                fcmTokenAndAlarmSettingDto.getFcmToken(),
-                                e.getMessage()
-                        );
-
                     }
+
+
                 }
             } else {
                 log.error( "푸시 알림 실패 해당 맴버 토큰 없음 / id = {}", notificationDto.getTargetMemberId());
